build: move tool preview presentation to cli ui
