+++ docs/parity/PARITY_DASHBOARD.md
@@ -1,14 +1,14 @@
 # Parity Dashboard
 
-_Generated from source artifacts: `2026-06-23T11:00:17.533132+00:00`_
+_Generated from source artifacts: `2026-06-23T11:49:00.474768+00:00`_
 
 ## Snapshot
 
 - Upstream target: `upstream/main` @ `5ecf3bf0e0726b8b33682bb5c3aad9679b7b5be4`
-- Workstream snapshot generated: `2026-06-23T04:15:52-06:00`
+- Workstream snapshot generated: `2026-06-23T05:17:48-06:00`
 - Parity matrix generated: `2026-06-12T11:29:13.798398+00:00`
 - Queue snapshot generated: `2026-06-16T23:55:38.405701+00:00`
-- Proof snapshot generated: `2026-06-23T11:00:17.533132+00:00`
+- Proof snapshot generated: `2026-06-23T11:49:00.474768+00:00`
 
 ## Gate Status
