+++ docs/parity/global-parity-proof.json
@@ -260,7 +260,7 @@
       "unverified_cases": 0
     }
   },
-  "generated_at_utc": "2026-06-23T11:00:17.533132+00:00",
+  "generated_at_utc": "2026-06-23T11:49:00.474768+00:00",
   "gpar_completion": {
     "GPAR-01": true,
     "GPAR-02": true,
