+++ docs/parity/global-parity-proof.md
@@ -1,6 +1,6 @@
 # Global Parity Proof
 
-Generated: `2026-06-23T11:00:17.533132+00:00`
+Generated: `2026-06-23T11:49:00.474768+00:00`
 
 ## Gate Status
