+++ docs/parity/test-intent-mapping.json
@@ -1,5 +1,5 @@
 {
-  "generated_at_utc": "2026-06-23T10:59:20.156630+00:00",
+  "generated_at_utc": "2026-06-23T11:48:53.426974+00:00",
   "intent_unit": "domain_intent",
   "intents": [
     {
