+++ docs/parity/workstream-status.json
@@ -1,6 +1,6 @@
 {
-  "generated_at_utc": "2026-06-23T04:15:52-06:00",
-  "head_sha": "3920d29c8217942c688847e760479f2af97c019d",
+  "generated_at_utc": "2026-06-23T05:17:48-06:00",
+  "head_sha": "f708494a083f916d02ab4ce056dc2cc2233f242c",
   "states": {
     "complete": 7
   },
