+++ docs/parity/workstream-status.md
@@ -1,6 +1,6 @@
 # Workstream Status
 
-- Local HEAD: `3920d29c8217942c688847e760479f2af97c019d`
+- Local HEAD: `f708494a083f916d02ab4ce056dc2cc2233f242c`
 - Upstream: `upstream/main` (`5ecf3bf0e0726b8b33682bb5c3aad9679b7b5be4`)
 
 | Workstream | Title | State |
