+++ scripts/audit-cargo-build-surface.sh
@@ -123,10 +123,10 @@ echo "## Interpretation"
 echo
 echo "- \`hermes-cli\` is currently the invalidation root for wrappers, the main runtime binary, TUI/clipboard UI dependencies, cron, ACP, MCP, tools, skills, and telemetry."
 echo "- Gateway adapter features are now behind explicit \`gateway-*\` CLI features. Default builds keep \`gateway-adapters-all\` for behavior compatibility, while \`--no-default-features\` checks compile the gateway core without adapter modules."
-echo "- Slash-command rendering, autocomplete ranking, alias canonicalization, and command catalog presentation now have a narrow dependency-free home in \`hermes-cli-ui\`."
+echo "- Slash-command rendering, autocomplete ranking, alias canonicalization, command catalog presentation, and gateway/tool preview rendering now have a narrow presentation home in \`hermes-cli-ui\`."
 echo "- Runtime platform tool-planning policy now has a narrower home in \`hermes-tool-planning\`; \`hermes-cli::platform_toolsets\` remains as an explicit compatibility re-export."
 echo "- Provider/auth routing now has a narrower home in \`hermes-provider-runtime\`; agent configuration, query-mode provider/model/env/tool policy, model remediation, noninteractive agent-loop wiring, reply extraction, and runtime prompt reformulation policy now have a narrower home in \`hermes-app-runtime\`."
 echo "- Source/governance parity now lives in \`hermes-source-parity-tests\`, so command-contract checks avoid \`hermes-cli\`, \`clap\`, fixture-harness runtime crates, and protocol stack crates."
 echo "- Protocol differential parity now lives in \`hermes-protocol-parity-tests\`, isolating ACP/MCP/gateway/tool dependencies to tests that need them."
-echo "- The next low-risk presentation split candidate is pure gateway/tool preview rendering; broader terminal UI, checklist, theme, and clipboard extraction should wait for a dedicated TUI crate boundary."
+echo "- Broader terminal UI, checklist, theme, and clipboard extraction should wait for a dedicated TUI crate boundary."
 echo "- Parity tests that only validate provider, auth, app-runtime, or command contracts should keep moving to narrower crates instead of pulling the full CLI binary surface."
